feat(leetcode): Add Python script for scraping
The leetcode.py file in scripts/ goes through the process of identifying
all questions on LeetCode while also assigning each one a tag
corresponding to its related data structure or algorithm.

The current method of determining which tags go to which questions is
done quite crudely; I've gone ahead and copy and pasted the contents of
each tag's page in a .txt file and then parsed that afterwords.

Might want to update this process to automate the retrieval of such
information.