fix(config): Fix config form submission
The issue with the form submission was not anything with Go or gRPC but
rather the way I set up the actual HTML form. Submissions were
redirecting back to config/ without any id and thus, we could not access
Firestore.

The template could still use some cleanup in terms of validation and
user feedback but it seems to work properly now.